fix piperead() with lazy sbrk
Assuming xv6 wants to support applications read()ing from a pipe into
lazy-sbrk-allocated memory, the kernel must be careful about errors
from copyout() during piperead(), because those errors might not be the
application's fault.  So, piperead() must return the number of bytes
already copied to user, or -1 if the very first byte caused a copyout()
error, to avoid returning 0 signaling that the pipe is closed.